+++ src/main.py
@@ -14,25 +14,32 @@ async def main():
 
     # coordinates from 0,0 to 10,10
     coordinator = RegionalCoordinator(
-        f"regional_coordinator1@{DEFAULT_HOST}", AGENT_PASSWORD, 0, 10, 0, 10, parking_agents_jids=["parking100"]
+        f"regional_coordinator1@{DEFAULT_HOST}",
+        AGENT_PASSWORD,
+        0,
+        10,
+        0,
+        10,
+        parking_agents_jids=[f"parking100@{DEFAULT_HOST}"],
     )
     await coordinator.start()
     print("Regional Coordinator started")
 
-    # coordinates from 10,0 to 20,10
-    coordinator = RegionalCoordinator(f"regional_coordinator2@{DEFAULT_HOST}", AGENT_PASSWORD, 10, 20, 0, 10)
-    await coordinator.start()
-    print("Regional Coordinator started")
+    # commented to simplify tests
+    # # coordinates from 10,0 to 20,10
+    # coordinator = RegionalCoordinator(f"regional_coordinator2@{DEFAULT_HOST}", AGENT_PASSWORD, 10, 20, 0, 10)
+    # await coordinator.start()
+    # print("Regional Coordinator started")
 
-    # coordinates from 0,10 to 10,20
-    coordinator = RegionalCoordinator(f"regional_coordinator3@{DEFAULT_HOST}", AGENT_PASSWORD, 0, 10, 10, 20)
-    await coordinator.start()
-    print("Regional Coordinator started")
+    # # coordinates from 0,10 to 10,20
+    # coordinator = RegionalCoordinator(f"regional_coordinator3@{DEFAULT_HOST}", AGENT_PASSWORD, 0, 10, 10, 20)
+    # await coordinator.start()
+    # print("Regional Coordinator started")
 
-    # coordinates from 10,10 to 20,20
-    coordinator = RegionalCoordinator(f"regional_coordinator4@{DEFAULT_HOST}", AGENT_PASSWORD, 10, 20, 10, 20)
-    await coordinator.start()
-    print("Regional Coordinator started")
+    # # coordinates from 10,10 to 20,20
+    # coordinator = RegionalCoordinator(f"regional_coordinator4@{DEFAULT_HOST}", AGENT_PASSWORD, 10, 20, 10, 20)
+    # await coordinator.start()
+    # print("Regional Coordinator started")
 
 
 if __name__ == "__main__":